requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.0.0
antlr4-python3-runtime==4.8
anyio==3.7.1
apptools==5.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru @ file:///home/conda/feedstock_root/build_artifacts/async-lru_1690563019058/work
async-timeout==4.0.2
attrs==23.1.0
Babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1677767029043/work
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
chumpy==0.70
click==8.1.7
coloredlogs==15.0.1
comm==0.1.3
configobj==5.0.8
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
envisage==7.0.3
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-pasta==0.2.0
h5py==3.9.0
humanfriendly==10.0
hydra-core==1.1.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
importlib-metadata @ file:///croot/importlib-metadata_1678997070253/work
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1688248289187/work
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp @ file:///home/conda/feedstock_root/build_artifacts/jupyter-lsp-meta_1685453365113/work/jupyter-lsp
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_1692015883666/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1690205927615/work
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
llvmlite==0.40.1
lxml @ file:///croot/lxml_1679646450407/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1691139191280/work
notebook_shim==0.2.3
numba==0.57.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020413938/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
overrides==7.3.1
packaging @ file:///croot/packaging_1678965309396/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
pooch==1.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyDeprecate==0.3.2
pyface==8.0.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110947380/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-ranger==0.1.1
pytorch-toolbelt==0.5.0
pytorchcv==0.0.65
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
pyvista==0.41.0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
scooby==0.7.2
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smplx==0.1.26
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.5
tinycss2==1.2.1
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
torch==2.1.1
torch-optimizer==0.1.0
torchgeometry==0.1.2
torchmetrics==1.0.1
torchvision==0.10.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.4.1
traitsui==8.0.0
trame==2.5.2
trame-client==2.10.0
trame-components==2.1.1
trame-deckgl==2.0.2
trame-markdown==2.0.2
trame-matplotlib==2.0.2
trame-plotly==2.1.1
trame-rca==0.3.1
trame-router==2.0.2
trame-server==2.11.7
trame-simput==2.3.2
trame-vega==2.0.3
trame-vtk==2.5.7
trame-vuetify==2.3.1
triton==2.1.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
vtk==9.2.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
wrapt==1.15.0
wslink==1.11.1
xmltodict==0.13.0
yarl==1.9.2
zipp @ file:///croot/zipp_1672387121353/work
